Clean up and coment, also move a funcution to modules
